Cleaning up some type assertion digging
