+++ cmd/kitgen/flatlayout.go
@@ -1,0 +1,33 @@
+package main
+
+type flat struct{}
+
+func (f flat) transformAST(ctx *sourceContext) (files, error) {
+	root := &ast.File{
+		Name:  ctx.pkg,
+		Decls: []ast.Decl{},
+	}
+
+	addImports(root, ctx)
+
+	for _, iface := range ctx.interfaces { //only one...
+		addStubStruct(root, iface)
+
+		for _, meth := range iface.methods {
+			addMethod(root, iface, meth)
+			addRequestStruct(root, meth)
+			addResponseStruct(root, meth)
+			addEndpointMaker(root, iface, meth)
+		}
+
+		addEndpointsStruct(root, iface)
+		addHTTPHandler(root, iface)
+
+		for _, meth := range iface.methods {
+			addDecoder(root, meth)
+			addEncoder(root, meth)
+		}
+	}
+
+	return formatNodes(outputTree{"gokit.go": root})
+}
